added minor changes to navigator package and added noteboogs for 2-point and FLI sampling
